JBEHAVE-1241 Moved stories containing pending steps.
