+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/AndStep.java
@@ -1,0 +1,6 @@
+package org.jbehave.examples.core.stories.failing;
+
+import org.jbehave.examples.core.CoreStory;
+
+public class AndStep extends CoreStory {
+}
