+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/and_step.story
@@ -1,0 +1,6 @@
+Scenario: An initial And step should be marked as pending as there is not previous step
+
+!-- What is this And of?  JBehave treats as pending 
+And the wind blows
+!-- Look Ma' - I can also use keywords in scenario title and step comments!
+
