+++ examples/core/src/main/java/org/jbehave/examples/core/stories/failing/examples_table_failing.story
